% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.update}
\alias{radiant.update}
\alias{radiant.update-package}
\alias{radiant.update}
\title{radiant.update}
\usage{
radiant.update()
}
\description{
radiant.update

Update Radiant
}
\examples{
\dontrun{
radiant.update::radiant.update()
}

}
