% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{remove_user_packages}
\alias{remove_user_packages}
\title{Remove packages that are installed in a user's local packages directory}
\usage{
remove_user_packages()
}
\description{
Remove packages that are installed in a user's local packages directory
}
