% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{user_packages}
\alias{user_packages}
\title{Check which packages are installed in a user's local packages directory}
\usage{
user_packages()
}
\description{
Check which packages are installed in a user's local packages directory
}
