% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.update}
\alias{radiant.update}
\alias{radiant.update-package}
\alias{radiant.update}
\title{radiant.update}
\usage{
radiant.update(lib.loc = .libPaths()[1],
  repos = "https://radiant-rstats.github.io/minicran/", dev = FALSE, type)
}
\arguments{
\item{lib.loc}{Library to install packages in (see .libPaths())}

\item{repos}{Repo to update from (default is the radiant repo on GitHub)}

\item{dev}{If TRUE, add the radiant development repo to the repo list}

\item{type}{Package type ("binary" or "source"). If missing, will try infer from OS type (i.e., "source" for linux, else "binary")}
}
\description{
radiant.update

Update Radiant
}
\examples{
\dontrun{
radiant.update::radiant.update()
}

}
