% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.update}
\alias{radiant.update}
\title{radiant.update}
\usage{
radiant.update(lib.loc = .libPaths()[1], repos = "", type, options = "")
}
\arguments{
\item{lib.loc}{Library to install packages in (see .libPaths())}

\item{repos}{Repo to update from (default is the radiant repo on GitHub)}

\item{type}{Package type ("binary" or "source"). If missing, will try infer from OS type (i.e., "source" for linux, else "binary")}

\item{options}{Install options as a string to pass to install.packages (e.g., '--no-lock')}
}
\description{
Update radiant and dependencies to the latest (supported) versions
}
\examples{
\dontrun{
radiant.update::radiant.update()
}

}
