% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{sync_packages}
\alias{sync_packages}
\title{Sync packages to the version on the miniCRAN repo}
\usage{
sync_packages(lib.loc = .libPaths()[1], type)
}
\arguments{
\item{lib.loc}{Library to install packages in}

\item{type}{Package type}
}
\description{
Sync packages to the version on the miniCRAN repo
}
\examples{
\dontrun{
radiant.update::sync_packages()
}

}
